% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-to-recurrent.R
\name{ct_to_recurrent}
\alias{ct_to_recurrent}
\title{Convert camera trap data into a recurrent event format}
\usage{
ct_to_recurrent(
  data,
  primary,
  secondary,
  survey_end_date,
  tertiary = NULL,
  datetime_var = "DateTime",
  species_var = "Species",
  site_var = "Site",
  survey_duration = 10
)
}
\arguments{
\item{data}{A dataframe containing the site, the timestamp and the species information}

\item{primary}{A pool of one or several species names considered to affect the secondary species}

\item{secondary}{Name of the species affected by the primary (only one name allowed).}

\item{survey_end_date}{Date of study end (e.g. "01-01-2001")}

\item{datetime_var}{Name of the variable that contains date time information.}

\item{species_var}{Name of the variable that contains species information.}

\item{site_var}{Name of the variable that contains Site ID.}

\item{survey_duration}{Maximum duration of the survey (in days, e.g. "7")}
}
\value{
A dataframe containing:
\describe{
\item{Site}{Camera-trap site ID}
\item{ID}{Survey ID}
\item{primary}{primary species starting the survey}
\item{DateTime_Primary}{Timestamp of the start of the survey (i.e. of the primary species observation)}
\item{DateTime}{Timestamp of the recurrent event (i.e. of the secondary species observation)}
\item{t.start}{Time since the start of the survey}
\item{t.stop}{Time since the last event, which can be the primary species or a previous secondary species}
\item{event}{Binary, 1 = recurrent event, 0 = censoring event}
\item{enum}{Event number within the survey}
\item{Event_type}{Either the recurrent event, giving the name of the secondary species, or a censoring event and the reason of censoring}
\item{DateTime_Primary}{Timestamp of survey start (primary observation)}
\item{...}{Any covariate attached to the dataset}
}
}
\description{
Takes a camera trap dataset as input (site ID, timestamp and species name) and return a recurrent event dataset.
The function creates a survey after each observation of the primary species for a defined maximum duration.
Within each survey, all the observations of the secondary species are converted into recurrent events.
The returned dataframe can be directly used for recurrent event analysis.
}
